+++ pkg/deployment/reconcile/plan_builder.go
@@ -193,6 +193,15 @@ func createPlan(ctx context.Context, log zerolog.Logger, apiObject k8sutil.APIOb
 		}
 	}
 
+	// Update status
+	if plan.IsEmpty() {
+		plan = pb.Apply(createEncryptionKeyStatusUpdate)
+	}
+
+	if plan.IsEmpty() {
+		plan = pb.Apply(createTLSStatusUpdate)
+	}
+
 	// Check for scale up/down
 	if plan.IsEmpty() {
 		plan = pb.Apply(createScaleMemeberPlan)
@@ -208,19 +217,26 @@ func createPlan(ctx context.Context, log zerolog.Logger, apiObject k8sutil.APIOb
 		plan = pb.Apply(createEncryptionKey)
 	}
 
-	// Check for the need to rotate TLS certificate of a members
 	if plan.IsEmpty() {
-		plan = pb.Apply(createRotateTLSServerCertificatePlan)
+		plan = pb.Apply(createCARenewalPlan)
 	}
+
+	if plan.IsEmpty() {
+		plan = pb.Apply(createCAAppendPlan)
+	}
+
+	if plan.IsEmpty() {
+		plan = pb.Apply(createKeyfileRenewalPlan)
+	}
+
+	// Check for the need to rotate TLS certificate of a members
+	//if plan.IsEmpty() {
+	//	plan = pb.Apply(createRotateTLSServerCertificatePlan)
+	//}
 
 	// Check for changes storage classes or requirements
 	if plan.IsEmpty() {
 		plan = pb.Apply(createRotateServerStoragePlan)
-	}
-
-	// Check for the need to rotate TLS CA certificate and all members
-	if plan.IsEmpty() {
-		plan = pb.Apply(createRotateTLSCAPlan)
 	}
 
 	if plan.IsEmpty() {
@@ -233,6 +249,10 @@ func createPlan(ctx context.Context, log zerolog.Logger, apiObject k8sutil.APIOb
 
 	if plan.IsEmpty() {
 		plan = pb.Apply(cleanEncryptionKey)
+	}
+
+	if plan.IsEmpty() {
+		plan = pb.Apply(createCACleanPlan)
 	}
 
 	// Return plan